double to int

+++ OrganicEncounter/model/Stats.java
@@ -11,24 +11,24 @@ package model;
 */
 
 public class Stats {
-    private double health;
-    private double rep;
-    private double money;
+    private int health;
+    private int rep;
+    private int money;
 
-    public Stats(double health, double rep, double money) {
+    public Stats(int health, int rep, int money) {
         this.health = health;
         this.rep = rep;
         this.money = money;
     }
 
     // getters 
-    public double getHealth() { return health; }
-    public double getRep() { return rep; }
-    public double getMoney() { return money; }
+    public int getHealth() { return health; }
+    public int getRep() { return rep; }
+    public int getMoney() { return money; }
 
     // modifies the stats to store
-    public void modifyHealth(double amount) { health += amount; }
-    public void modifyRep(double amount) { rep += amount; }
-    public void modifyMoney(double amount) { money += amount; }
+    public void modifyHealth(int amount) { health += amount; }
+    public void modifyRep(int amount) { rep += amount; }
+    public void modifyMoney(int amount) { money += amount; }
 
 }
